requirements.txt
langchain-openai
langchain
python-dotenv
langchain-community
streamlit